+++ R/add_boundary_penalties.R
@@ -120,12 +120,17 @@ NULL
 #'
 #' Note that when the problem objective is to maximize some measure of
 #' benefit and not minimize some measure of cost, the term \eqn{p} is
-#' replaced with \eqn{-p}.
+#' replaced with \eqn{-p}. Additionally, to linearize the problem,
+#' the expression \eqn{X_{iz} \times X_{jy}}{Xiz * Xjy} is modeled using
+#' a set of continuous variables (bounded between 0 and 1) based on
+#' Beyer *et al.* (2016).
 #'
 #' @return An updated [problem()] object with the penalties added to it.
 #'
 #' @seealso
 #' See [penalties] for an overview of all functions for adding penalties.
+#' Additionally, see [add_neighbor_penalties()] for a penalty that
+#' can reduce spatial fragmentation and has faster solver run times.
 #'
 #' @family penalties
 #'
@@ -232,7 +237,7 @@ NULL
 #' # solve problems
 #' s2 <- list(solve(p5), solve(p6), solve(p7), solve(p8), solve(p9), solve(p10))
 #'
-#' # convert to category layers for visualization
+#' #convert to category layers for visualization
 #' s2 <- terra::rast(lapply(s2, category_layer))
 #' names(s2) <- c(
 #'   "basic solution", "within zone clumping (low)",