+++ R/add_max_utility_objective.R
@@ -45,7 +45,7 @@ NULL
 #' Here, \eqn{x_i}{xi} is the [decisions] variable (e.g.,
 #' specifying whether planning unit \eqn{i}{i} has been selected (1) or not
 #' (0)), \eqn{r_{ij}}{rij} is the amount of feature \eqn{j}{j} in planning
-#' unit \eqn{i}{i}, \eqn{A_j}{Aj} is the amount of feature \eqn{j}{j}
+#' unit \eqn{i}{i}, \eqn{a_j}{aj} is the amount of feature \eqn{j}{j}
 #' represented in in the solution, and \eqn{w_j}{wj} is the weight for
 #' feature \eqn{j}{j} (defaults to 1 for all features; see
 #' [add_feature_weights()]